addressing MarcHDF comments
