+++ Reactions/ReactorCvodeUtils.H
@@ -1,7 +1,7 @@
 #ifndef _REACTORUTILSCVODE_H_
 #define _REACTORUTILSCVODE_H_
 #include "ReactorTypes.H"
-#include "ReactorCvode.H"
+
 #include <AMReX.H>
 #include <AMReX_Print.H>
 #include <AMReX_REAL.H>
@@ -10,9 +10,86 @@
 #include <sunmatrix/sunmatrix_sparse.h>
 #include "ReactorUtils.H"
 
+#ifdef AMREX_USE_CUDA
+#include <cusolverSp.h>
+#endif
+
+#ifdef PELE_USE_KLU
+#include "klu.h"
+#include <sunlinsol/sunlinsol_klu.h>
+#endif
+
 namespace pele {
 namespace physics {
 namespace reactions {
+
+struct CVODEUserData
+{
+  amrex::Real dt_save;     // Internal cvode dt holder
+  int ncells;              // Number of cells in solve
+  int verbose;             // Verbose
+  int maxOrder;            // CVODE max order
+  int reactor_type;        // Either HP (LM) or UV (C)
+  int analytical_jacobian; // Analytical Jacobian 'On' flag
+  int solve_type;          // Type of linear solve for Newton direction
+  int precond_type;        // Type of preconditioner (if iterative solve)
+  int NNZ; // Number of non-zero entry for sparse representations
+  amrex::Real* rhoe_init = nullptr; // Initial energy (rhoE for C, rhoH for LM)
+  amrex::Real* rhoesrc_ext = nullptr; // External energy forcing
+  amrex::Real* rYsrc_ext = nullptr;   // External species forcing
+  int* mask =
+    nullptr; // Masking tagging cells where integration should not be perfomed
+  int* FCunt = nullptr; // Number of RHS evaluations (not used on GPU)
+  amrex::Real gamma;    // System Chem. jacobian coefficient
+  int nbBlocks;         // GPU kernel launch parameter
+  int nbThreads;        // GPU kernel launch parameter
+
+#ifdef AMREX_USE_GPU
+  // Matrix data
+  int* csr_row_count_h;   // Host CSR sparse mat
+  int* csr_col_index_h;   // Host CSR sparse mat
+  int* csr_row_count_d;   // Device CSR sparse mat
+  int* csr_col_index_d;   // Device CSR sparse mat
+  amrex::Real* csr_val_d; // Syst. Jac. CSR sparse mat values
+  amrex::Real* csr_jac_d; // Chem. Jac. CSR sparse mat values
+
+#ifdef AMREX_USE_CUDA
+  // Batched QR preconditioner data
+  void* buffer_qr = NULL;    // cuSolver internal work arrays
+  csrqrInfo_t info;          // QR solve info
+  cusparseMatDescr_t descrA; // Sparse mat. descriptor
+  cusolverSpHandle_t cusolverHandle;
+  cusparseHandle_t cuSPHandle;
+#endif
+  amrex::gpuStream_t stream;
+
+#else
+  bool FirstTimePrecond;
+  amrex::Real** Jdata = NULL;
+  // Sparse direct solve data
+  // Ptrs to CSR/CSC matrix for each submatrices (cells)
+  SUNMatrix* PS;
+  int** rowVals = NULL;
+  int** rowPtrs = NULL;
+  int** colPtrs = NULL;
+  int** colVals = NULL;
+  int* indx = NULL;
+  amrex::Real** JSPSmat = NULL;
+#ifdef PELE_USE_KLU
+  klu_common* Common;
+  klu_symbolic** Symbolic;
+  klu_numeric** Numeric;
+#endif
+  // Matrix data for denseSimpleAjac preconditioner
+  amrex::Real**** Jbd;
+  amrex::Real**** P;
+  sunindextype*** pivot;
+  // Custom direct linear solve: a CSR SUNMatrix and ptrs
+  SUNMatrix PSc;
+  int* colVals_c;
+  int* rowPtrs_c;
+#endif
+};
 
 namespace cvode {
