+++ Testing/Exec/Make.PelePhysics
@@ -132,7 +132,7 @@ endif
 
 include $(AMREX_HOME)/Tools/GNUMake/Make.rules
 
-#$(objForExecs): TPL
+$(objForExecs): TPL
 
 ifeq ($(USE_CUDA),TRUE)
    CXXFLAGS+=-Xptxas --disable-optimizer-constants
